requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aiosignal==1.3.1
anndata==0.10.8
array_api_compat==1.8
async-timeout==4.0.3
attrs==24.2.0
blessed==1.20.0
build==1.2.2.post1
CacheControl==0.14.1
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
chex==0.1.86
cleo==2.1.0
click==8.1.7
cmake==3.30.2
contextlib2==21.6.0
contourpy==1.2.1
crashtest==0.4.1
cryptography==44.0.0
cycler==0.12.1
datasets==2.21.0
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.9
docker-pycreds==0.4.0
docrep==0.3.2
dulwich==0.21.7
einops==0.8.0
etils==1.5.2
exceptiongroup==1.2.2
fastjsonschema==2.21.1
filelock==3.15.4
flash-attn==1.0.4
flax==0.8.5
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
get-annotations==0.1.2
gitdb==4.0.11
GitPython==3.1.43
gpustat==1.1.1
h5py==3.11.0
huggingface-hub==0.24.5
idna==3.7
igraph==0.11.6
importlib_metadata==8.2.0
importlib_resources==6.4.3
installer==0.7.0
jaraco.classes==3.4.0
jax==0.4.30
jaxlib==0.4.30
jeepney==0.8.0
Jinja2==3.1.4
joblib==1.4.2
keyring==24.3.1
kiwisolver==1.4.5
leidenalg==0.10.2
lightning==2.1.4
lightning-utilities==0.11.6
lit==18.1.8
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-collections==0.1.1
ml-dtypes==0.4.0
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.0.8
mudata==0.2.4
multidict==6.0.5
multipledispatch==1.0.0
multiprocess==0.70.16
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
numba==0.60.0
numpy==1.26.4
numpyro==0.15.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.560.30
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opt-einsum==3.3.0
optax==0.2.3
orbax-checkpoint==0.5.23
packaging==24.1
pandas==2.2.2
patsy==0.5.6
pexpect==4.9.0
pillow==10.4.0
pkginfo==1.12.0
platformdirs==4.2.2
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
protobuf==5.27.3
psutil==6.0.0
ptyprocess==0.7.0
pyarrow==17.0.0
pycparser==2.22
pydot==3.0.1
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.1.2
pyproject_hooks==1.2.0
pyro-api==0.1.2
pyro-ppl==1.9.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.3.3
pytz==2024.1
PyYAML==6.0.2
RapidFuzz==3.10.1
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
safetensors==0.4.4
scanpy==1.9.8
scib==1.1.5
scikit-learn==1.5.1
scikit-misc==0.3.1
scipy==1.13.1
scvi-tools==1.1.5
seaborn==0.13.2
SecretStorage==3.3.3
sentry-sdk==2.13.0
session-info==1.0.0
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
statsmodels==0.14.2
stdlib-list==0.10.0
sympy==1.13.2
tensorstore==0.1.64
texttable==1.7.0
threadpoolctl==3.5.0
tokenizers==0.19.1
tomli==2.2.1
tomlkit==0.13.2
toolz==0.12.1
torch==2.0.1
torchaudio==2.0.2
torchdata==0.6.1
torchmetrics==1.4.1
torchtext==0.15.2
torchvision==0.15.2
tqdm==4.66.5
transformers==4.44.0
triton==2.0.0
trove-classifiers==2024.10.21.16
typing_extensions==4.12.2
tzdata==2024.1
umap-learn==0.5.6
urllib3==2.2.2
virtualenv==20.28.0
wandb==0.17.7
wcwidth==0.2.13
wrapt==1.16.0
xxhash==3.5.0
yarl==1.9.4
zipp==3.20.0
